+++ src/sdk/PnP.Core.Test/SharePoint/ListItemTests.cs
@@ -280,6 +280,53 @@ namespace PnP.Core.Test.SharePoint
         }
 
         [TestMethod]
+        public async Task BulkAddListItemsWithBadFieldNameTestDontThrowOnError()
+        {
+            //TestCommon.Instance.Mocking = false;
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite))
+            {
+                var listTitle = TestCommon.GetPnPSdkTestAssetName("BulkAddListItemsWithBadFieldNameTest");
+                var list = await context.Web.Lists.AddAsync(listTitle, ListTemplateType.GenericList);
+
+                IField myField = await list.Fields.AddTextAsync("MetaInfo", new FieldTextOptions()
+                {
+                    Group = "Custom Fields",
+                    AddToDefaultView = true,
+                });
+
+                Assert.IsTrue(myField.InternalName != "MetaInfo");
+
+                List<Dictionary<string, object>> propertiesToUpdate = new List<Dictionary<string, object>>();
+
+                var prop = new Dictionary<string, object>
+                {
+                    { "Title", "My title 1" },
+                    { "MetaInfo", "abc" }
+                };
+                propertiesToUpdate.Add(prop);
+
+                var prop2 = new Dictionary<string, object>
+                {
+                    { "Title", "My title 2" },
+                    { "MetaInfo", "abc" }
+                };
+                propertiesToUpdate.Add(prop2);
+
+                foreach (var propItem in propertiesToUpdate)
+                {
+                    await list.Items.AddBatchAsync(propItem);
+                }
+                
+                var errors = await context.ExecuteAsync(false);
+
+                Assert.IsTrue(errors.Count == 2);
+
+                await list.DeleteAsync();
+            }
+        }
+
+
+        [TestMethod]
         public async Task VerifyUserFieldsInRepetiveUpdates()
         {
             //TestCommon.Instance.Mocking = false;